Rename LEVINE_NICOLAS_FINAL_PROJECT.ipynb to NYC Car Crash Data Analysis.ipynb
In this project, I was tasked with finding an accredited data source, cleaning the data, and then doing some analysis on it. Within the cells of my code, I have provided notes to show what my thought process was. Furthermore, there is a summary of my findings at the bottom.